requirements.txt
flask==3.1.0
gunicorn==23.0.0
requests==2.32.4
urllib3==2.6.3
matplotlib==3.10.1
pillow==12.1.1
numpy==1.26.4
